+++ webui/react/src/pages/TrialDetails/TrialDetailsOverview.tsx
@@ -1,18 +1,34 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
+import { MINIMUM_PAGE_SIZE } from 'components/Table';
 import useStorage from 'hooks/useStorage';
 import TrialInfoBox from 'pages/TrialDetails/TrialInfoBox';
 import { parseUrl } from 'routes/utils';
-import { ExperimentBase, MetricName, MetricType, TrialDetails } from 'types';
+import { ApiSorter } from 'services/types';
+import { ExperimentBase, MetricName, MetricType, Pagination, TrialDetails } from 'types';
 import { extractMetricNames } from 'utils/trial';
 
 import TrialChart from './TrialChart';
 import TrialDetailsWorkloads from './TrialDetailsWorkloads';
 
+const defaultSorter: ApiSorter = {
+  descend: true,
+  key: 'batches',
+};
+
+export enum TrialInfoFilter {
+  Checkpoint = 'Has Checkpoint',
+  Validation = 'Has Validation',
+  CheckpointOrValidation = 'Has Checkpoint or Validation',
+}
+
 const URL_ALL = 'all';
 
 const STORAGE_CHART_METRICS_KEY = 'metrics/chart';
 const STORAGE_PATH = 'trial-detail';
+const STORAGE_LIMIT_KEY = 'limit';
+const STORAGE_CHECKPOINT_VALIDATION_KEY = 'checkpoint-validation';
+const STORAGE_SORTER_KEY = 'sorter';
 
 export interface Props {
   experiment: ExperimentBase;
@@ -45,6 +61,17 @@ const TrialDetailsOverview: React.FC<Props> = ({ experiment, trial }: Props) => 
   const [ metrics, setMetrics ] = useState<MetricName[]>(
     storage.getWithDefault(storageChartMetricsKey || '', initMetric ? [ initMetric ] : []),
   );
+  const initFilter = storage.getWithDefault(
+    STORAGE_CHECKPOINT_VALIDATION_KEY,
+    TrialInfoFilter.CheckpointOrValidation,
+  );
+  const initLimit = storage.getWithDefault(STORAGE_LIMIT_KEY, MINIMUM_PAGE_SIZE);
+  const initSorter = storage.getWithDefault(STORAGE_SORTER_KEY, { ...defaultSorter });
+  const [ sorter, setSorter ] = useState<ApiSorter>(initSorter);
+  const [ pagination, setPagination ] = useState<Pagination>(
+    { limit: initLimit, offset: 0 },
+  );
+  const [ showFilter, setShowFilter ] = useState(initFilter);
 
   /*
    * When filters changes update the page URL.
@@ -54,6 +81,21 @@ const TrialDetailsOverview: React.FC<Props> = ({ experiment, trial }: Props) => 
 
     const searchParams = new URLSearchParams;
     const url = parseUrl(window.location.href);
+
+    // limit
+    searchParams.append('limit', pagination.limit.toString());
+
+    // offset
+    searchParams.append('offset', pagination.offset.toString());
+
+    // sortDesc
+    searchParams.append('sortDesc', sorter.descend ? '1' : '0');
+
+    // sortKey
+    searchParams.append('sortKey', sorter.key || '');
+
+    // filters
+    searchParams.append('filter', showFilter ? showFilter : URL_ALL);
 
     // metrics
     if (metrics && metrics.length > 0) {
@@ -67,7 +109,7 @@ const TrialDetailsOverview: React.FC<Props> = ({ experiment, trial }: Props) => 
       '',
       url.origin + url.pathname + '?' + searchParams.toString(),
     );
-  }, [ isUrlParsed, metrics ]);
+  }, [ metrics, isUrlParsed, pagination, showFilter, sorter ]);
 
   /*
      * On first load: if filters are specified in URL, override default.
@@ -84,6 +126,30 @@ const TrialDetailsOverview: React.FC<Props> = ({ experiment, trial }: Props) => 
 
     const urlSearchParams = url.searchParams;
 
+    // limit
+    const limit = urlSearchParams.get('limit');
+    if (limit != null && !isNaN(parseInt(limit))) {
+      pagination.limit = parseInt(limit);
+    }
+
+    // offset
+    const offset = urlSearchParams.get('offset');
+    if (offset != null && !isNaN(parseInt(offset))) {
+      pagination.offset = parseInt(offset);
+    }
+
+    // sortDesc
+    const sortDesc = urlSearchParams.get('sortDesc');
+    if (sortDesc != null) {
+      sorter.descend = (sortDesc === '1');
+    }
+
+    // filter
+    const filter = urlSearchParams.get('filter');
+    if (filter != null) {
+      setShowFilter(filter as TrialInfoFilter);
+    }
+
     // metrics
     const visibleMetrics = urlSearchParams.getAll('metric');
     let metrics = defaultMetrics;
@@ -93,9 +159,19 @@ const TrialDetailsOverview: React.FC<Props> = ({ experiment, trial }: Props) => 
         return { name: splitMetric[0], type: splitMetric[1] as MetricType };
       }));
     }
+
+    // sortKey
+    const sortKey = urlSearchParams.get('sortKey');
+    if (sortKey != null &&
+      [ 'batches', 'state', ...metrics.map(metric => metric.name) ].includes(sortKey)) {
+      sorter.key = sortKey;
+    }
+
     setIsUrlParsed(true);
+    setPagination(pagination);
+    setSorter(sorter);
     setMetrics(metrics);
-  }, [ defaultMetrics, isUrlParsed, metrics ]);
+  }, [ defaultMetrics, isUrlParsed, pagination, showFilter, sorter, metrics ]);
 
   const handleMetricChange = useCallback((value: MetricName[]) => {
     setMetrics(value);
@@ -128,6 +204,12 @@ const TrialDetailsOverview: React.FC<Props> = ({ experiment, trial }: Props) => 
         handleMetricChange={handleMetricChange}
         metricNames={metricNames}
         metrics={metrics}
+        pagination={pagination}
+        setPagination={setPagination}
+        setShowFilter={setShowFilter}
+        setSorter={setSorter}
+        showFilter={showFilter}
+        sorter={sorter}
         trial={trial} />
     </>
   );
